requirements.txt
PyQt5
dnspython